+++ chunkflow/operators/__init__.py
@@ -3,7 +3,6 @@
 """
 from .cloud_watch import CloudWatchOperator
 from .create_chunk import CreateChunkOperator
-from .create_thumbnail import CreateThumbnailOperator
 from .crop_margin import CropMarginOperator
 from .cutout import CutoutOperator
 from .inference import InferenceOperator
@@ -12,6 +11,5 @@ from .neuroglancer_view import NeuroglancerViewOperator
 from .normalize_section_contrast import NormalizeSectionContrastOperator
 from .read_file import ReadFileOperator
 from .save import SaveOperator
-from .upload_log import UploadLogOperator
 from .view import ViewOperator
 from .write_h5 import WriteH5Operator