+++ chunkflow/operators/save.py
@@ -1,13 +1,30 @@
+import os
+import json
 import numpy as np
 from cloudvolume import CloudVolume 
+from cloudvolume.lib import Vec, Bbox
+from cloudvolume.storage import Storage
+
+from chunkflow.igneous.tasks import downsample_and_upload
 
 from .operator_base import OperatorBase
 
 
 class SaveOperator(OperatorBase):
-    def __init__(self, volume_path: str, mip: int, 
+    def __init__(self, volume_path: str, mip: int,
+                 upload_log: bool=True,
+                 create_thumbnail: bool=False,
+                 nproc: int=0,
                  verbose: bool=True, name: str='save'):
         super().__init__(name=name)
+        if nproc < 0:
+            nproc = True
+        elif nproc == 0:
+            nproc = False
+        
+        self.upload_log = upload_log
+        self.create_thumbnail = create_thumbnail
+        self.mip = mip
         self.verbose = verbose
         self.volume = CloudVolume(
             volume_path,
@@ -15,10 +32,65 @@ class SaveOperator(OperatorBase):
             bounded=False,
             autocrop=True,
             mip=mip,
+            parallel=nproc,
             progress=verbose)
+
+        if upload_log:
+            log_path = os.path.join(volume_path, 'log')
+            self.log_storage = Storage(log_path)
+
+        if create_thumbnail:
+            self.thumbnail_volume = CloudVolume(
+                os.path.join(volume_path, 'thumbnail'),
+                compress='gzip',
+                fill_missing=True,
+                bounded=False,
+                autocrop=True,
+                mip=mip,
+                progress=verbose)
  
-    def __call__(self, chunk):
+ 
+    def __call__(self, chunk, log=None, output_bbox=None):
         chunk_slices = chunk.slices 
         # transpose czyx to xyzc order
         arr = np.transpose(chunk)
         self.volume[chunk_slices[::-1]] = arr
+
+        if self.upload_log:
+            self._upload_log(log, output_bbox)
+
+        if self.create_thumbnail:
+            self._create_thumbnail(chunk)
+
+    def _create_thumbnail(self, chunk):
+        if self.verbose:
+            print('creating thumbnail...')
+        # only use the last channel, it is the Z affinity
+        # if this is affinitymap
+        image = chunk[-1, :, :, :]
+        if np.issubdtype(image.dtype, np.floating):
+            image = (image * 255).astype(np.uint8)
+
+        # transpose to xyzc
+        image = np.transpose(image)
+        image_bbox = Bbox.from_slices(chunk.slices[::-1][:3])
+        
+        downsample_and_upload(
+            image,
+            image_bbox,
+            self.thumbnail_volume,
+            Vec(*(image.shape)),
+            mip=self.mip,
+            axis='z',
+            skip_first=True,
+            only_last_mip=True)
+ 
+    def _upload_log(self, log, output_bbox):
+        assert log
+        assert isinstance(output_bbox, Bbox)
+        # write to google cloud storage 
+        self.log_storage.put_file(
+            file_path=output_bbox.to_filename() + '.json',
+            content=json.dumps(log),
+            content_type='application/json')
+